fix test, un-comment it
this test wasn't waiting for an async to finish before
using its "return value."

and it wasn't waiting at the end, either.
so if you ran the test in isolation, it passed,
but if you ran it as part of a suite, later on
it would async'ly throw an exception

fixed those.

split this long test into two smaller tests so
I don't have to think so hard to figure out what's
going on.