fix flaky test (#54)
one of these tests failed once: it endorsed a token, then "found"
that token again with getAccessTokens; it compared these two copies
of the token by equality. This mostly passed, but it failed just now:
the two token protos had a couple of signatures in a different order.
Semantically, these tokens are "the same", but proto isEqual: check
didn't like it.

So change that test to check for equal .id_p instead of the whole structure.
And change some other checks, too. Just because they've been passing
so far doesn't mean they won't go flaky later.