sample code += use access token that maybe does not have allAccounts (#83)
so far, our redeem-access-token sample code shows how to use a token hat has allAccounts. But if the token-grantor replaces that allAccounts with something more specific, e.g., edits the permissions in our iPhone app, that sample code fails to fetch any data. This new sample code shows how to use allAccounts if granted else use more-specific permissions.

bonus: fix old replace-access sample mistake: it was trying to set new config by calling methods on old config

bonus: fix a test that was comparing a pointer to string "10" being above 0, rather than converting that "10" to an int. it worked in that context, but was confusing.